HUGE COMMIT: add all models from master branch
This commit adds all the models from the master branch. NEEDS CLEANING
Also solves the issue with the iris_labyrinth.world where the arrows
won't show due to a model naming collision.
This commit also solves the trees population of the
quadrotor_2_hangar.world not showing due to the same problem.

This commit includes minor fixes.